+++ tqsdk/__init__.py
@@ -3,7 +3,8 @@
 __author__ = 'chengzhi'
 name = "tqsdk"
 
-from tqsdk.api import TqApi, TqAccount
+from tqsdk.api import TqApi
+from tqsdk.account import TqAccount
 from tqsdk.channel import TqChan
 from tqsdk.backtest import TqBacktest, TqReplay
 from tqsdk.exceptions import BacktestFinished
